LayoutTests/fast/css-grid-layout/grid-item-order-in-content-sized-columns-resolution-expected.txt

PASS grid-template-columns is 50px 20px after grid row swap.
PASS grid-template-columns is 70px 50px after grid row swap.
PASS grid-template-columns is 50px 20px after grid row swap.
PASS grid-template-columns is 50px 20px after grid row swap.
PASS grid-template-columns is 70px 0px after grid row swap.
PASS grid-template-columns is 50px 20px after grid row swap.
PASS grid-template-columns is 50px 20px after grid row swap.
PASS grid-template-columns is 50px 20px after grid row swap.
PASS successfullyParsed is true

TEST COMPLETE
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
XXXXX
XXX XXX
